test: when messages expire expect show next message in local queue
https://linear.app/customerio/issue/MBL-381/after-message-expires-replace-in-app-message-with-next-in-queue-if-one

This ticket has already been implemented and no SDK code needs to be modified. To verify this, this PR adds more automated tests for scenarios the SDK could get in for messages expiring.

Testing
* There are 2 test functions that test messages expiring and displaying new messages later on.
* I tested this logic in UIKit sample app. Testing in-app message expiration is difficult because [messages have a minimum of 1 hour of expiration](https://customerio.slack.com/archives/C036FTE6UJ0/p1717051120718159?thread_ts=1717003069.505409&cid=C036FTE6UJ0).

Instead of waiting for 1 hour, I simulated a message expiring by modifying the SDK code and running the sample app.