test: given expired message not yet shown expect to not show message
https://linear.app/customerio/issue/MBL-381/after-message-expires-do-not-display-it-in-app-if-it-has-not-already

As a marketer, I don't want messages that are no longer relevant to be displayed in an app. If a message expires, I expect it to not be displayed in an app if it has not been shown yet.

The logic of this ticket is already implemented in the SDK. This PR is only adding test cases that are missing for the acceptance criteria of the ticket.

From the acceptance criteria of the ticket, there are 2 test cases that need tested:
* Given a message expires and is already being shown, expect continue showing it until message is closed. (test function `test_expiration_givenDisplayedMessageExpires_expectContinueShowingMessageUntilClose` already covers this)
* Given a message expires and not yet shown, expect to not show that message. This PR adds this test case.