working implementation.
APN works as expected. appdelegate function gets called as expected.

for FCM, the feature works with the FCM SDK, but the original appdelegate didregister function is not called. I think it's because it's a swiftui app. we are sending the request back to original imp of swiftui appdelegate and not uikit one.