remove setting frame in sample app
Now that the SwiftUI implementation is complete, I decided to check and see if setting the frame() is required or not. From testing, it's not needed. the width of the View is automatically taking the full width of the parent.

Our docs are going to mention that it's up to the developer to set the width of the View, but does not give a suggestion for how to do that. To match our docs, I suggest we remove this un-needed line from our sample app.